Implement parameter compilation during the first processing phase.
This commit implements the Action.parameters() method which allows
actions to inject key/value pairs into the parameter list during
the first processing phase, when it hasn't yet been determined which
actions will be handling the request. The changes are used by
URIFilter and WerkzeugMapFilter to pass parameters from the URL to
other actions.