Add parameter list to Action construction.
This commit adds the parameter list (actually parameter dict) as an
argument to the handles(), __new__(), and __init__() methods of each
Action and to the handle() method of the Action metaclass. This allows
an Action subclass to decide whether or not to handle the request based
on not only the request itself, but the contents of the parameter list,
as assembled by previously tested actions. Currently only the filters
URIFilter and WerkzeugMapFilter place anything into the parameter list
via the new mechanism. (All other modifications to the parameter list
are still done through generate() as before.)